:bug: 메인 피드 리뷰 사항 반영
