+++ feature/main/src/main/kotlin/com/hm/picplz/ui/screen/main/MainSideEffect.kt
@@ -7,5 +7,7 @@ sealed interface MainSideEffect {
 
     data class NavigateToPhotographerDetail(val photographerId: Long) : MainSideEffect
 
+    data object ShowReportUnavailable : MainSideEffect
+
     data object NavigateToDev : MainSideEffect
 }
