+++ feature/main/src/main/res/values/strings.xml
@@ -206,4 +206,5 @@
         <item>필름 카메라</item>
     </string-array>
     <string name="main_feed_load_more_retry">작가를 더 불러오지 못했어요. 다시 시도</string>
+    <string name="main_report_unavailable">신고 기능을 준비 중이에요</string>
 </resources>
